Removed memcpy, need to cudafy BVs

+++ grid.h
@@ -1,11 +1,11 @@
 #ifndef OTJ_GRID_HEADER
 #define OTJ_GRID_HEADER 1
 typedef struct grid_t {
-	float* internal_storage;
+	double* internal_storage;
 	int len_x;
 	int len_y;
 } Grid;
 
-#define grid_size(g) sizeof(float) * g.len_x * g.len_y
+#define grid_size(g) sizeof(double) * g.len_x * g.len_y
 #define grid_element(g,i,j) g.internal_storage[(i * g.len_x) + (j - 1)]
 #endif
